+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Uliweb Change Log
 * Change `generic/template_files/common/views.py` to `views.py.tmpl`
 * Add GenericReference value `(Model, object_id)` support
 * Fix pyini save process
+* Fix `_create_kwargs` quote characters bug, and change the function to `to_attrs`
 
 0.1.6 Version
 -----------------